Add NotesInputField component & handler
Seperated the notes text area into a seperate component, with the state
being managed within App.js for easy submission once all fields are
filled out. Handler is handleInputChange - which can be used for
multiple input fields. This will help if the application grows in the
future as 'handleInputChange' can be used for all inputs.